deployment/cdk/requirements.txt
aws-cdk-lib==2.166.0
boto3
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha
cdklabs.generative-ai-cdk-constructs==0.1.287